feat: introduce Block and validation rules with mining functionality

+++ api/src/main/kotlin/dtos/Transaction.kt
@@ -3,6 +3,6 @@ package api.dtos
 data class Transaction(
     val from: String,
     val to: String,
-    val amount: Float,
+    val amount: Long,
     val signature: String,
 )
